Added debugger functionality: breakpoints from CLI, state output
git-svn-id: https://mash16.googlecode.com/svn/trunk@91 f32040c0-604e-ac78-71ca-de615562ae79

+++ src/core/cpu.h
@@ -78,7 +78,7 @@ typedef struct flags
 typedef struct cpu_meta
 {
     long cycles;
-    double target_cycles;
+    long target_cycles;
     int wait_vblnk;
 
 } cpu_meta;
